requirements.txt
streamlit>=1.30.0
zhipuai>=2.0.0
mistune>=3.0.0
beautifulsoup4>=4.12.0
readability-lxml>=0.8.1
requests>=2.31.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
lxml>=4.9.0
